+++ e2e/token-vault-suites/teardown.ts
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) 2021 - 2026 Ping Identity Corporation. All rights reserved.
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
 export default () => {
   console.log('tests finished');
 };