+++ packages/javascript-sdk/src/device-client/mock-data/msw-mock-data.ts
@@ -1,3 +1,12 @@
+/*
+ * @forgerock/javascript-sdk
+ *
+ * msw-mock-data.ts
+ *
+ * Copyright (c) 2025 - 2026 Ping Identity Corporation. All rights reserved.
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
 import { GeneralResponse } from '../services/index.js';
 import type {
   OathResponse,